+++ packages/read/src/index.ts
@@ -13,7 +13,7 @@ async function filterChangesetsSinceRef(
     cwd: changesetBase,
     ref: sinceRef,
   });
-  const newHashes = newChangesets.map((c) => c.split("/")[1]);
+  const newHashes = newChangesets.map((c) => c.split("/").pop());
 
   return changesets.filter((dir) => newHashes.includes(dir));
 }